Issue#1Lesson#4: Upgrade functions Created structure for classes
